Correção limitar para impedir criação de contas duplicadas com tipos diferentes
